Fix : merging of the BasketController and DvdBasketController

+++ second/src/main/java/com/dvdstore/second/dtos/DvdBasketGetDTO.java
@@ -1,15 +1,13 @@
 package com.dvdstore.second.dtos;
 
-import com.dvdstore.second.domains.BasketRepositoryModelPostgreSQL;
-import com.dvdstore.second.repositories.BasketRepository;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
 public class DvdBasketGetDTO {
-    private Integer id;
-    private Integer dvdId;
-    private Integer quantity;
-    private Integer price;
+    private int id;
+    private int dvdId;
+    private int quantity;
+    private float price;
 }
